Added width and height properties to the Figure.
Now able to resize plot, and have it immediately update the Figure settings.

Adding a wave to a table now puts it to the left of the clicked on wave, not to the right (or at 0).

Tick lengths now default to 8 (major) and 3 (minor).

Tick direction can now be in, out, or both.

Remove dist. btwn tick and label, as it is not used by pygraphene.

Added log plotting.

Removed point, pixel, y-all, pentagon, hexagon, diamond, thin diamond marker types.